refactor: move utilities/password and factory to packages
