+++ apps/web/__tests__/pages/api/auth/auth.test.ts
@@ -1,5 +1,5 @@
 import '__mocks__/tokens';
-import { create } from '__tests__/factory';
+import { create } from '@linen/factory';
 import { v4 } from 'uuid';
 import { testApiHandler } from 'next-test-api-route-handler';
 import handler from 'pages/api/auth/[[...slug]]';
