+++ apps/web/components/layout/PageLayout/Header/index.test.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import Header from '.';
 import { render } from '@testing-library/react';
-import { build } from '__tests__/factory';
+import { build } from '@linen/factory';
 import { Mode } from '@linen/hooks/mode';
 
 jest.mock('utilities/auth/react', () => ({
